requirements.txt
frida-tools
PyQt6
pyinstaller